requirements.txt
google-generativeai
flask
flask_cors
markdown